v5 first commit

+++ modules/System/database/migrations/2023_05_08_043740_create_system_dictionary_values.php
@@ -13,22 +13,25 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('system_dictionary_values', function (Blueprint $table) {
-            $table->id();
-            $table->integer('dic_id')->comment('字典ID');
-            $table->string('label')->comment('值名称');
-            $table->tinyInteger('value')->comment('对应值');
-            $table->integer('sort')->default(0)->comment('排序');
-            $table->tinyInteger('status')->default(1)->comment('状态 1 正常 2 禁用');
-            $table->string('description', 1000)->comment('描述')->default('');
-            $table->creatorId();
-            $table->createdAt();
-            $table->updatedAt();
-            $table->deletedAt();
+        if (! Schema::hasTable('system_dictionary_values')) {
+            Schema::create('system_dictionary_values', function (Blueprint $table) {
+                $table->id();
+                $table->integer('dic_id')->comment('字典ID');
+                $table->string('label')->comment('值名称');
+                $table->string('key')->comment('值KEY');
+                $table->tinyInteger('value')->comment('对应值');
+                $table->integer('sort')->default(0)->comment('排序');
+                $table->tinyInteger('status')->default(1)->comment('状态 1 正常 2 禁用');
+                $table->string('description', 1000)->comment('描述')->default('');
+                $table->creatorId();
+                $table->createdAt();
+                $table->updatedAt();
+                $table->deletedAt();
 
-            $table->engine='InnoDB';
-            $table->comment('字典对应值');
-        });
+                $table->engine = 'InnoDB';
+                $table->comment('字典对应值');
+            });
+        }
     }
 
     /**
